Expect test_321_ResultFetchArraysLobsXMLTypes to fail
The code does an SQLGetData, binding as SQL_CLOB_LOCATOR, but
DB2 Connect CLI returns invalid conversion to bind a CLOB to a
locator for some reason:
07006[IBM][CLI Driver] CLI0102E  Invalid conversion. SQLSTATE=07006

The BLOB path works, using SQLGetData with SQL_BLOB_LOCATOR,
so I'm not sure what's wrong with CLOBs. It seems like a bug, but I
can't get it to work and the test predates the CI setup, so just
expect it to fail for now.