var_dump has differing precision between PHP 7 and 8
This is probably due to the locale-neutral float conversion.
var_dump will show an extra 3 digits of precisions varying, but
var_export seems to always show the remaining 3 digits. Many of the
other conversion functions could have been used instead if we did
not want to show the last three digits.