Fix pconnect keepalive check not freeing statement fully
The code to free statement resources frees everything held by the
structure, but not the statement resource itself, since I believe
these can be allocated outside of new_struct. The free code seems
to call the special free function then PHP efree, so we'll do the
same in the pconnect check.