Change DBCLOB+GetSubString to work from first null instead of backwards
GetSubString with an oversized buffer (like from i5_dbcs_alloc) can
sometimes add garbage to the end of the buffer when using DBCLOB locators
with SQL_C_CHAR outputs. The old code seemed to scan backwards, which is
bad if it hits the garbage. As such, try to scan from the first null
and then clear out spaces only after getting everything after the null.
Since the PASE driver implies DBCLOB is for strings and not binary data,
this should be OK?