Workaround for possible bug caused by SQL/CLI PTF
It seems db2_result was returning garbage at the end of strings;
an unknown PTF seemed to be causing this. My guess is something
related to Unicode conversion based on the PTFs of the time. It
manifests returning more data than requested, usually around twice
the buffer size. As such, just allocate twice as much and truncate.
Not ideal, but we're kinda forced to do this if we want to avoid a
buffer overflow.

It's affected one customer, but now it has affected another, so I'm
moving this into git.