fetch system naming mode at runtime (PASE)
This is similar to what PDO_IBM does. The benefit is for applications
like the Toolkit for generating queries at runtime, when they get a
resource passed in (since they didn't construct it and don't know).

Note that like PDO_IBM, other options are possible, but not yet
implemented. Comments are made for the possibility though.

Does not affect LUW.