Include password in persistent hash key
Matches behaviour of other database extensions, and removes some
surprise that the extension doesn't check for a valid password when
reusing a connection. Of course, that must be the password the
connection was created with, but that's the same for the others too.

Fixes GH-18

XXX: Would like to include options, but not sure how to do this cleanly.