Fix garbage in column names on IBM i
A user reported an issue where the column names can have garbage at the
end in some situations; in this case, it seems to be using QP2SHELL
while also have the CCSID set to 65535. It seems SQL/CLI doesn't null
terminate the names when this happens. Arguably a bug in SQL/CLI, and
arguably on the user as QP2SHELL requires you to set up the environment
yourself, but unfortunately CCSID 65535 and QP2SHELL usage are common.

This resolves the issue by truncating with `name_length`, which is still
set correctly.