Try to use out_length for bound fetched values
I have no idea why it's using strlen in the first place, but if SQL/CLI
returns junk in the buffer, then the length might be correct. This
matches what we did with column names on IBM i.

The SQL_BINARY case is more puzzling, but we can take a look at it later
if this works for strings. Why is it ifdef'd to use strlen/out_length
depending on platform?